refactor: simplify type resolution
