+++ packages/typescript-estree/src/ts-estree/ts-nodes.ts
@@ -11,21 +11,24 @@ declare module 'typescript' {
 
 export type TSToken = ts.Token<ts.SyntaxKind>;
 
-/**
- * List of nodes that are skipped
- */
+// TODO: drop after TSESTreeToTSNode refactor
 export type TSNodeSkipped =
+  // List of nodes that are skipped
   | ts.OmittedExpression
-  | ts.ComputedPropertyName // we are skipping this node
-  | ts.CaseBlock
-  | ts.HeritageClause
-  | ts.NamedImports
-  | ts.NamedExports
-  | ts.TemplateSpan
+  | ts.ComputedPropertyName
   | ts.ParenthesizedExpression
   | ts.Token<ts.SyntaxKind.ImportKeyword>;
 
-export type TSNodeUnsupported =
+// TODO: drop after TSESTreeToTSNode refactor
+export type TSNodeUnused =
+  // Nodes that are skipped when generating code
+  | ts.HeritageClause
+  | ts.CaseBlock
+  | ts.NamedImports
+  | ts.TemplateSpan
+  | ts.NamedExports
+
+  // Those nodes can't be generated by ast
   | ts.PartiallyEmittedExpression
   | ts.SyntheticExpression
   | ts.NotEmittedStatement
@@ -73,7 +76,7 @@ export type TSNodeUnsupported =
 // TODO: replace with code that retrieves all nodes that have kind
 // export type TSNode2 = Extract<typeof ts[keyof typeof ts], { kind: any }>
 export type TSNode =
-  | TSNodeUnsupported
+  | TSNodeUnused
   | TSNodeSkipped
   | ts.Modifier
   | ts.Identifier
@@ -218,27 +221,3 @@ export type TSNode =
   | ts.SourceFile
   | ts.TemplateLiteralTypeNode;
 
-export type TSNodeSupported = Exclude<TSNode, TSNodeUnsupported>;
-export type TSNodeExpression = Exclude<
-  Extract<TSNodeSupported, ts.Expression>,
-  // manual fixes
-  | ts.OmittedExpression // there is advanced handling for this node in type guards
-  | ts.Token<ts.SyntaxKind.ImportKeyword> // this node can be generated only in call expression
->;
-export type TSNodeStatement = Extract<TSNodeSupported, ts.Statement>;
-export type TSNodeTypeNode = Exclude<
-  Extract<TSNodeSupported, ts.TypeNode>,
-  // manual fixes
-  ts.ExpressionWithTypeArguments
->;
-export type TSNodeTypeElement = Extract<TSNodeSupported, ts.TypeElement>;
-export type TSNodeClassElement = Extract<TSNodeSupported, ts.ClassElement>;
-
-export type TSNodeConvertable =
-  | TSNode
-  | ts.Expression
-  | ts.Statement
-  | ts.TypeNode
-  | ts.TypeElement
-  | ts.ClassElement
-  | undefined;
